+++ Sofa/framework/Simulation/Core/test/MappingGraph_test.cpp
@@ -21,7 +21,7 @@
 ******************************************************************************/
 #include <sofa/testing/BaseTest.h>
 #include <sofa/simulation/MappingGraph.h>
-#include <sofa/simulation/MappingGraph2.h>
+#include <sofa/simulation/MappingGraph.h>
 #include <sofa/core/MechanicalParams.h>
 #include <sofa/simulation/Node.h>
 #include <sofa/simpleapi/SimpleApi.h>
@@ -31,7 +31,7 @@
 
 TEST(MappingGraph, noBuild)
 {
-    const sofa::simulation::MappingGraph2 graph;
+    const sofa::simulation::MappingGraph graph;
 
     EXPECT_FALSE(graph.isBuilt());
     EXPECT_EQ(graph.getRootNode(), nullptr);
@@ -47,7 +47,7 @@ TEST(MappingGraph, noBuild)
 
 TEST(MappingGraph, nullRootNode)
 {
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(nullptr);
 
     EXPECT_FALSE(graph.isBuilt());
@@ -66,7 +66,7 @@ TEST(MappingGraph, emptyRootNode)
 {
     const sofa::simulation::Node::SPtr root = sofa::core::objectmodel::New<sofa::simulation::Node>();
 
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(root.get());
 
     EXPECT_TRUE(graph.isBuilt());
@@ -89,7 +89,7 @@ TEST(MappingGraph, oneMechanicalObject)
     root->addObject(mstate);
     mstate->resize(10);
 
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(root.get());
 
     EXPECT_TRUE(graph.isBuilt());
@@ -114,7 +114,7 @@ TEST(MappingGraph, twoMechanicalObject)
     root->addObject(mstate2);
     mstate2->resize(2);
 
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(root.get());
 
     EXPECT_TRUE(graph.isBuilt());
@@ -147,7 +147,7 @@ TEST(MappingGraph, oneMapping)
     mapping->setFrom(mstate1.get());
     mapping->setTo(mstate2.get());
 
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(root.get());
 
     EXPECT_TRUE(graph.isBuilt());
@@ -208,7 +208,7 @@ TEST(MappingGraph, diamondMapping)
         {"input", "@/left/left @/right/right"}, {"output", "@bottom"}
     });
 
-    sofa::simulation::MappingGraph2 graph;
+    sofa::simulation::MappingGraph graph;
     graph.build(root.get());
 
     EXPECT_TRUE(graph.isBuilt());
